replace ng8 with ng15 project
