+++ scripts/test-performance.sh
@@ -7,11 +7,12 @@ cd $DIR/..
 npm run build
 
 # Test Angular build size
-cd $DIR/../test/performance/test-angular
+
+cd $DIR/../test/performance/angular-app
 npm install
 
-mkdir -p $DIR/../test/performance/test-angular/node_modules/@alfresco/js-api
-cp -R $DIR/../dist/* $DIR/../test/performance/test-angular/node_modules/@alfresco/js-api
+mkdir -p $DIR/../test/performance/angular-app/node_modules/@alfresco/js-api
+cp -R $DIR/../dist/* $DIR/../test/performance/angular-app/node_modules/@alfresco/js-api
 
 npm run build:prod || exit 1
