+++ test/performance/angular-app/src/app/app.component.ts
@@ -1,20 +1,3 @@
-/*!
- * @license
- * Copyright © 2005-2023 Hyland Software, Inc. and its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 import { Component, OnInit } from '@angular/core';
 import { AlfrescoApi, DiscoveryEntry, DiscoveryApi } from '@alfresco/js-api';
 
@@ -24,7 +7,6 @@ import { AlfrescoApi, DiscoveryEntry, DiscoveryApi } from '@alfresco/js-api';
   styleUrls: ['./app.component.css']
 })
 export class AppComponent implements OnInit {
-
   alfrescoApi: AlfrescoApi;
   title = 'My test performance app';
   ecmVers: any = {};
@@ -48,5 +30,4 @@ export class AppComponent implements OnInit {
       this.ecmVers = ecmVers;
     });
   }
-
 }
