--- test/performance/test-angular/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "js-api-angular",
-  "version": "0.0.0",
-  "scripts": {
-    "build": "ng build",
-    "build:prod": "ng build --prod && npm run bundlesize-check",
-    "bundlesize-check": "bundlesize"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^8.2.11",
-    "@angular/common": "^8.2.11",
-    "@angular/compiler": "^8.2.11",
-    "@angular/core": "^8.2.11",
-    "@angular/platform-browser": "^8.2.11",
-    "@angular/platform-browser-dynamic": "^8.2.11",
-    "core-js": "^2.5.4",
-    "rxjs": "^6.5.3",
-    "zone.js": "~0.9.1"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "~0.803.12",
-    "@angular/cli": "^8.3.12",
-    "@angular/compiler-cli": "^8.2.11",
-    "@angular/language-service": "^8.2.11",
-    "@types/node": "^12.11.2",
-    "bundlesize": "^0.18.0",
-    "typescript": "~3.5.3"
-  },
-  "bundlesize": [
-    {
-      "path": "./dist/js-api-angular/main.js",
-      "maxSize": "90 kb"
-    }
-  ]
-}
