--- test/performance/test-angular/proxy.conf.js
@@ -1,45 +1,0 @@
-/*!
- * @license
- * Copyright © 2005-2023 Hyland Software, Inc. and its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-var PROXY_HOST_ADF = process.env.PROXY_HOST_ADF;
-var HOST = process.env.URL_HOST_ADF;
-
-module.exports = {
-    "/alfresco": {
-        "target": "http://" + (PROXY_HOST_ADF || HOST),
-        "secure": false,
-        "pathRewrite": {
-            "^/alfresco/alfresco": ""
-        },
-        "changeOrigin": true,
-        // workaround for REPO-2260
-        onProxyRes: function (proxyRes, req, res) {
-            const header = proxyRes.headers['www-authenticate'];
-            if (header && header.startsWith('Basic')) {
-                proxyRes.headers['www-authenticate'] = 'x' + header;
-            }
-        }
-    },
-    "/activiti-app": {
-        "target": "http://" + (PROXY_HOST_ADF || HOST),
-        "secure": false,
-        "pathRewrite": {
-            "^/activiti-app/activiti-app": ""
-        },
-        "changeOrigin": true
-    }
-};
