--- test/performance/test-angular/src/app/app.component.html
@@ -1,3 +1,0 @@
-<div style="text-align:center">
-</div>
-
