--- test/performance/test-angular/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "baseUrl": "./",
-    "outDir": "./dist/out-tsc",
-    "sourceMap": true,
-    "declaration": false,
-    "module": "es2015",
-    "moduleResolution": "node",
-    "skipLibCheck": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-    "allowSyntheticDefaultImports": true,
-    "sourceRoot": "../src",
-    "target": "es5",
-    "typeRoots": [
-      "node_modules/@types"
-    ],
-
-    "lib": [
-      "es2017",
-      "dom"
-    ]
-  }
-}
